# services/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GleanWise",
    "description": "Local-first answer engine. `POST /query` then stream `GET /stream/{id}` (SSE).\nOpenAI-compatible endpoint: `POST /v1/chat/completions`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/search": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Web search (no LLM)",
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/fetch": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Fetch and clean one page",
        "operationId": "fetch_page_fetch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/retrieve": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Passages from the local index",
        "operationId": "retrieve_passages_retrieve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrieveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proxy/favicon": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Privacy-preserving favicon (fetched server-side)",
        "operationId": "favicon_proxy_favicon_get",
        "parameters": [
          {
            "name": "domain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 253,
              "title": "Domain"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/proxy/image": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Fetch a remote image server-side so the browser never contacts the origin",
        "description": "Answers can embed images, and a prompt-injected page could point one at an attacker's server\nto leak data in the URL. Routing every image through here (SSRF-guarded, raster types only,\nsize-capped) removes both the tracking and the exfiltration channel.",
        "operationId": "proxy_image_proxy_image_get",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 2048,
              "title": "Url"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FetchRequest": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 4000,
            "title": "Url"
          },
          "max_chars": {
            "type": "integer",
            "maximum": 200000.0,
            "minimum": 200.0,
            "title": "Max Chars",
            "default": 20000
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "FetchRequest"
      },
      "FetchResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "final_url": {
            "type": "string",
            "title": "Final Url"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated",
            "default": false
          },
          "from_cache": {
            "type": "boolean",
            "title": "From Cache",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "url",
          "final_url",
          "text"
        ],
        "title": "FetchResponse"
      },
      "Focus": {
        "type": "string",
        "enum": [
          "general",
          "academic",
          "news",
          "social"
        ],
        "title": "Focus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetrieveRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "limit": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 8
          },
          "urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "RetrieveRequest"
      },
      "RetrieveResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/RetrievedChunk"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "query",
          "chunks"
        ],
        "title": "RetrieveResponse"
      },
      "RetrievedChunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "heading_path": {
            "type": "string",
            "title": "Heading Path",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "url",
          "text",
          "score"
        ],
        "title": "RetrievedChunk"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "focus": {
            "$ref": "#/components/schemas/Focus",
            "default": "general"
          },
          "limit": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 10
          },
          "locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Results"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "query",
          "results"
        ],
        "title": "SearchResponse"
      },
...
      "Source": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "snippet": {
            "type": "string",
            "title": "Snippet",
            "default": ""
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "default": ""
          },
          "favicon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Favicon"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "engines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Engines"
          },
          "used_in_answer": {
            "type": "boolean",
            "title": "Used In Answer",
            "default": false
          },
          "state": {
            "type": "string",
            "enum": [
              "found",
              "reading",
              "read",
              "failed",
              "skipped"
            ],
            "title": "State",
            "default": "found"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url"
        ],
        "title": "Source"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}